+++ src/main/scala/dev/krysztal/are/common/component/TraitsMapComponent.scala
@@ -24,11 +24,10 @@
 
 package dev.krysztal.are.common.component
 
-import dev.krysztal.are.common.component.TraitsMapComponent.log
-import dev.krysztal.are.common.component.TraitsMapComponent.toKey
-import dev.krysztal.are.common.gene.GeneticSequence
 import net.minecraft.nbt.NbtCompound
+import net.minecraft.nbt.NbtElement
 import net.minecraft.registry.RegistryWrapper.WrapperLookup
+import net.minecraft.util.Identifier
 import net.minecraft.util.math.BlockPos
 import net.minecraft.world.chunk.Chunk
 import org.ladysnake.cca.api.v3.component.ComponentV3
@@ -36,8 +35,11 @@ import org.slf4j.LoggerFactory
 
 import scala.collection.immutable.HashMap
 
+private type GeneticTuple = (Identifier, NbtElement)
+
 trait TraitsMapComponent extends ComponentV3 {
-    val positionMap: HashMap[BlockPos, GeneticSequence] = HashMap()
+    val positionMap: HashMap[BlockPos, GeneticTuple] =
+        HashMap()
 }
 
 class TraitsMapComponentImpl(val chunk: Chunk) extends TraitsMapComponent {
@@ -45,12 +47,7 @@ class TraitsMapComponentImpl(val chunk: Chunk) extends TraitsMapComponent {
     override def readFromNbt(
         tag: NbtCompound,
         registryLookup: WrapperLookup
-    ): Unit = {
-        this.positionMap.foreach((p, g) => {
-            log.debug(s"Storage traits: $p -> $g")
-            tag.put(toKey(p), g.as())
-        })
-    }
+    ): Unit = {}
 
     override def writeToNbt(
         tag: NbtCompound,
@@ -61,10 +58,31 @@ class TraitsMapComponentImpl(val chunk: Chunk) extends TraitsMapComponent {
 object TraitsMapComponent {
     lazy val log = LoggerFactory.getLogger("TraitsMap")
 
-    def toKey(pos: BlockPos) = s"${pos.getX()};${pos.getY()};${pos.getZ()}"
+    def toKey(pos: BlockPos): String =
+        s"${pos.getX()},${pos.getY()},${pos.getZ()}"
 
-    def fromKey(str: String) = {
-        var pos = str.split(";")
+    def fromKey(str: String): BlockPos = {
+        var pos = str.split(",")
         BlockPos(pos(0).toInt, pos(1).toInt, pos(2).toInt)
+    }
+
+    def toTuple(nbt: NbtCompound): GeneticTuple = {
+        val gene = nbt.get("data");
+        val identifier = Identifier.of(nbt.getString("identifier"));
+
+        (identifier, gene)
+    }
+
+    def fromTuple(tup: GeneticTuple): NbtCompound = {
+        val (identifier, geneData) = tup
+
+        val compound = {
+            val data = NbtCompound()
+            data.putString("identifier", identifier.toString())
+            data.put("data", geneData)
+            data
+        }
+
+        compound
     }
 }