+++ src/main/scala/dev/krysztal/are/common/evaluator/Sugar.scala
@@ -22,36 +22,23 @@
  * SOFTWARE.
  */
 
-package dev.krysztal.are.common.gene
+package dev.krysztal.are.common.evaluator
 
-import com.mojang.serialization.MapCodec
+import dev.krysztal.are.common.gene.EnvironmentContext
 import dev.krysztal.are.foundation.as.AsNbtCompound
-import dev.krysztal.are.foundation.from.From
-import net.minecraft.nbt.NbtCompound
-import net.minecraft.util.math.BlockPos
-import net.minecraft.world.World
+import dev.krysztal.are.foundation.evaluator.Evaluator
+import dev.krysztal.are.foundation.extend.WorldExtension.temperature
 
-import scala.util.Try
+class Sugar extends Evaluator[EnvironmentContext] {
 
-case class EnvironmentContext(
-    val world: World,
-    val pos: BlockPos
-)
+    override def eval(ctx: EnvironmentContext): Class[? <: AsNbtCompound] = {
+        val lightLevel = ctx.world.getLightLevel(ctx.pos)
+        val temperature = ctx.world.temperature(ctx.pos)
+        ???
+    }
 
-/** Genetic sequence, the core of hybrid system.
-  *
-  * @author
-  *   KrysztalHuang <krysztal.huang@outlook.com>
-  */
-trait GeneticSequence extends AsNbtCompound, From[NbtCompound] {
-    def codec(): GeneticSequenceCodec[?]
-
-    def evaluate(ctx: EnvironmentContext): NbtCompound
-
-    def hybridize(
-        parentA: GeneticSequence,
-        parentB: GeneticSequence
-    ): Try[NbtCompound]
 }
 
-case class GeneticSequenceCodec[T <: GeneticSequence](codec: MapCodec[T])
+object Sugar {
+    lazy val INSTANCE = Sugar()
+}